[S8]  DAL Creation and refactoring
- Add method Load in orderDAL which returns list<orderDTO>
- Remove call to sp from the presentation layer and the load method of
orderDAL
- Create object orderLineDTO
- Create method Load in orderLineDAL which take a parameter OrderID, and
  call the usp_getorderline return a list<orderlineDTO>
- Link the orderviewline grid to Load of orderlineDAL as a datasource
- Create a load method on PersonDAL which take a personID and return back
  a personDTO
- Create ProductDAL to display products on placeorder screen
- ** Made start to vieworder screen intial load of grid which is commented
  out **

+++ training/SQL/BuildScript.sql
@@ -53,7 +53,7 @@ CREATE TABLE [product]
 (	productID int IDENTITY(1,1) Primary Key,
 	name VARCHAR (300) NOT NULL,
 	[description] VARCHAR (600) NOT NULL,
-	price FLOAT  Not Null);
+	price  numeric(18,2) Not Null);
 GO
 CREATE TABLE [personaddress]
 (	personaddressID int IDENTITY(1,1) Primary Key,
